# packages/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI CLI Syncer API
  description: |
    API for managing AI CLI tools (Claude, Cursor, Codex, Gemini, etc.) MCP configurations and Rules.
    This API enables synchronization of Model Context Protocol (MCP) servers and coding rules across multiple AI development tools.
  version: 0.1.0
  contact:
    name: AI CLI Syncer
  license:
    name: MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
...
paths:
...
  /rules/master:
    get:
      tags: [Rules]
      summary: Get master rules content
      operationId: getMasterRules
      responses:
        '200':
          description: Master rules content
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
    post:
      tags: [Rules]
      summary: Save master rules content
      operationId: saveMasterRules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveMasterRulesRequest'
      responses:
        '200':
          description: Rules saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
  /rules:
    get:
      tags: [Rules]
      summary: Get all rules
      operationId: getRulesList
      responses:
        '200':
          description: List of rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rule'
    post:
      tags: [Rules]
      summary: Create a new rule
      operationId: createRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRuleRequest'
      responses:
        '201':
          description: Rule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '400':
          $ref: '#/components/responses/ValidationError'
  /rules/{id}:
    put:
      tags: [Rules]
      summary: Update a rule
      operationId: updateRule
      parameters:
        - $ref: '#/components/parameters/RuleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRuleRequest'
      responses:
        '200':
          description: Rule updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Rules]
      summary: Delete a rule
      operationId: deleteRule
      parameters:
        - $ref: '#/components/parameters/RuleId'
      responses:
        '200':
          description: Rule deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          description: Cannot delete active rule
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    RuleId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Rule UUID
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    path:
                      type: string
                    message:
                      type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
  schemas:
...
    Rule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        content:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, content, isActive, createdAt, updatedAt]
...
    SaveMasterRulesRequest:
      type: object
      properties:
        content:
          type: string
      required: [content]
    CreateRuleRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        content:
          type: string
      required: [name, content]
    UpdateRuleRequest:
      type: object
      properties:
        content:
          type: string
        name:
          type: string
...